Fix the issue that PreviewViewMeteringPointFactory produces wrong coordinates when effect or stream sharing is enabled.
The PreviewviewMeteringPointFactory uses the PreviewRotationDegrees to transform the
coordinates. This PreviewRotationDegrees in TransformationInfo does't
reflect the sensor_orientation and the display orientation when effect
or stream sharing is enabled.

Solution :
  Use getSensorToBufferTransform from TransmationInfo instead to convert the view(x,y)
  to sensor (x,y) and also set the surfaceAspectRatio as the sensor size.

Bug: 345993685
RelNote: "Fixed the issue where invoking startFocusMetering with
PreviewView.getMeteringPointFactory does't get the correct sensor
coordinates when an effect is enabled or when 4 use cases are bound
(stream-sharing)."
Test: All tests in camera-view

Change-Id: I916c55d371a2c2e7de798b21080318591f099561