Supports ripple radius updates when bounds change on S+
On S+, if a new explicit radius is set before changing bounds, the new radius will be applied to an ongoing ripple.

Bug: b/360346196
Fixes: b/183019123
Test: RippleModifierNodeTest
Relnote: "Adds support for ripples in expanding components on Android S+, such as a card that expands on click. Previously the ripple would not fill up the new size but now it will expand to the new bounds"
Change-Id: If509acac87caa3f13aa2cf969f7bfade7da3e7f6