Removes unused parameters for Corners
in the earlier stages of the library, Corner has been added more parameters for experimentation purposes. after a couple of years now, it turns out that two of the parameters are unused. to simplify Corner, and therefore Feature, we remove the unused parameters vertex and roundedCenter. these changes only affect developers, as the constructor of Corner is internal.

Relnote: Removes unused parameters for Corners
Test: None
Change-Id: I4d316d27ddbb11a34b0fea8c67ed9ed8fab37ee3