Make compoundKeyOf iterative instead of recursive
This is a backport of the same change from the SlotTable rewrite CL
(aosp/2644758) where it yielded on average a 4% improvement in
SlotTableIntegrationBenchmark test cases at the time it was introduced.

I haven't taken new measurements for the cherry-pick, but anticipate
a similarly small but measurable improvement to editing operations.

Test: CompositeHashTests
Change-Id: I5817f443e766f85b874ccbdd11db16d20313f6df