Throw an error sooner for using destroyed NavController
Moving the LifecycleOwner of the NavController is a terminal state and
that NavController should not continue to be used.

This change ensure that calling setGraph on a destroyed NavController
will immediately throw an error.

NavController's that never had a lifecycle will work as expected.

RelNote: "Attempting to use a `NavController` that has been previously
`DESTROYED` will now cause an `IllegalStateException`."
Test: fixed nestedNavHostRestore test
Bug: 369616172

Change-Id: I520da520adf9f99d887e63c0255afe97ecefdab5