Temporarily disable building RISC-V artifacts
When users have a set of dependencies that contain some AARs that
support RISC-V and other AARs that don't, they can land in a state
where their release pipelines believe they are shipping RISC-V
artifacts, even if this wasn't their intention. This does not happen
if they *only* have AARs that support RISC-V.

We'll need to propagate changes in various systems to ensure that
RISC-V artifacts do not automatically attempt to be published and
produce reasonable warnings to explain how to work around this in
the future. For now, we'll stop building for this architecture to
allow stable releases of AndroidX to not break external clients.

We'll reland this change once we've addressed this issue and updated
our rollout strategy accordingly.

Bug: 366238650
Test: None
Change-Id: I14d1af41387d471c90ad44e01a5d5eee2d0e5346