Allow nested and fully overlapped ParagraphStyles
Previously we didn't allow overlapping paragraphs. However, it will make sense, especially for support of the bullet lists, to support nested or fully overlapped paragraphs.

Test: extensive unit test coverage in AnnotatedStringTest
Test: compose:ui:ui-text:test,  compose:ui:ui-text:cC, compose:f:f:cC
Relnote: "Updated how paragraphs are treated in AnnotatedString. Before you could only create non-overlapping paragraphs. Now the AnnotatedString allows fully-overlapping paragraphs which will be merged together, and nested paragraphs in which case the outer paragrapg will be splitted on bounds of the inner one, and the inner paragraph's style will be merged with the outer paragraph's style"

Change-Id: Ic9554c8201c430733543517dd697da3d5eace926