DatePicker colors with typography-defined colors
Fix the way we apply colors at the DatePicker and the
DateRangePicker when a `Typography` is configured to have TextStyles
with colors.

The change ensures that whatever is set at the DatePickerColors takes
presedence over any color that was set at the Typography's TextStyles.

Fixes: 347031394
Test: Added tests
Relnote: "DatePickerColors now correctly take precedence over any
conflicting colors defined at the theme's Typography text styles.
Also note that this update adjusts the `color` parameter's position in
the date picker functions and introduces a `contentColor` parameter
for customizing the header and title text colors."

Change-Id: I30d0307b11ba2e1a02535928ab4e4131100692a8