Allow R8 to remove debugging runtimeCheck calls
Adds a new debugRuntimeCheck function to the runtime that can be used
for Compose-developer facing checks. The goal is that R8 will strip out
more expensive checks that don't need to ship in production apps.
If the check would fail but was stripped, the program should still crash
later on, but with a less pretty exception.

Test: N/A
Change-Id: I8c441aa22f729e05651486265af9a2c0d9ee75b8