Update Camera2CameraController with latest status and retry
Update Camera2CameraController with latest camera availability status,
and attempt a restart when we receive a camera error but the camera has
already become available. The out-of-order scenario is possible with the
asynchronous nature of AvailabilityCallbacks and
CameraDevice.StateCallbacks, as well as how CameraPipe processes and
propagates camera closure states.

Camera2CameraController will be getting a minor scale local refactoring,
given the additional logic added to the current code. This will be
tracked by b/372258646.

Bug: 362902859
Test: CameraPipe sample app and CameraX integration tests.
Change-Id: I2c01fad855c0644a8d51d7248aa45f9da0920463