Add onRectChanged modifier and corresponding callback support to rect tracking
Relnote:"""
Added Modifier.onRectChanged API which is an API that allows one to subscribe to the root/window/screen-relative position and size of a
LayoutNode. The API solves many use cases that the existing onGloballyPositioned modifier does, however it does so with much less
overhead and the API comes with facilities to debounce and throttle the callback per what one's use case demands.
"""

Change-Id: I3c8fa250f561bbd846a9044118532ec439240efd