Add a LocalActivity composition local
Provides a fallback for being able to access Activity for APIs that
haven't been placed behind a testable interface. A common mistake when
faced with this situation is to directly cast Context, however this is not
always an activity. This CL provides a new LocalActivity computed composition local
that correctly walks up the context reference to find the current activity.

As compositionLocalWithComputedDefaultOf is only available in Compose runtime 1.7.0,
this required bumping the Compose runtime dependency.

Fixes: 283009666
Test: LocalActivityTest
Relnote: Add a `LocalActivity` composition local
Relnote: Activity Compose now depends on Compose Runtime 1.7.0
Change-Id: I7746aaaeb779d7af79867b4d9143b9d5c5545543