Optimize ObjectList
Get exception throws and string building out of the hot paths.
Also move array resize to a separate function while keeping
the size check inline. The next step will be to guarantee the
resizeStorage() function is never inline as the code it contains
is pretty large on aarch64. Note that creating the array and
calling copyInto() produces better code than using copyOf().

Relnote: N/A
Test: ObjectListTest
Change-Id: I336438c858583f11f831c3d4ebced25c95506f94