Add DEFAULT Embedding Search Metric Enum as a default enum.
Add DEFAULT Embedding Search Metric Enum as a default enum for
situations when there is no embedding set but we are returning the
embedding search metric used. This is used to represent that the default
set by SearchSpec will be used.

Relnote: Add DEFAULT Embedding Search Metric Enum as default enum
Test: Ran SearchSpecCtsTest and SearchSpecInternalTest to verify no breaking behavior.
Change-Id: I928abdb05f31caadedaa3b45e87fade4ba410da7