Check for whether java compile task will exist instead of throwing and catching
It would seem that we don't need to configure the Java plugin if it doesn't exist and the multiplatform project has no Java code. However the configureWithJavaPlugin does a bunch of other unrelated things that if not called here, don't get called anywhere and untangling that is outside the scope of this bug, but potentially worth looking at eventually.

Bug: 372454957
Test: ./development/validateRefactor.sh HEAD^
Change-Id: Ia9df610624582af5d864e1dbf047c3702d5a86e4