Apply under-exposure workaround to all TCL devices
This change resolves an under-exposure issue observed in the preview when binding with VideoCapture on TCL devices. The workaround enforces the use of TEMPLATE_PREVIEW to maintain consistency with previews without VideoCapture. This workaround is applied universally to all TCL devices and cameras to ensure a consistent user experience.

Relnote: "Fixed preview and video under-exposure issue on TCL devices when VideoCapture is enabled."

Bug: 359425774
Test: manually testing
Change-Id: Id7a64a83ed74ff2f189cd315ede03c090ff818a0