Add onRectChanged modifier and corresponding callback support to rect tracking
Fixes: 372765423
Fixes: 372757007
Fixes: 372994338

Relnote:"""
Added Modifier.onRectChanged API which is an API that allows one to subscribe to the root/window/screen-relative position and size of a
LayoutNode. The API solves many use cases that the existing onGloballyPositioned modifier does, however it does so with much less
overhead and the API comes with facilities to debounce and throttle the callback per what one's use case demands.
"""

Change-Id: Id28c7c6c52ce3b18c9ac8c83cde48111a312cfc9