fix NPE when the platform emits a null bt device
A Quality Assurance test surfaced a bug where the client app would
crash due to a NullPointerException on devices running older sdks. The
stack trace showed a NPE in the onCallAudioStateChanged callback.

Upon investigating, if the CallAudioState(activeBtDevice=null) is given
to the jetpack layer, this WILL cause a NPE when converting the
CallAudioState to CallEndpoints.  The code has been adjusted to prevent
this NPE.

Fixes: 372851034
Test: manual + unit
Change-Id: I784feab5303bdd04fba822f6e1c4da4cd9a6e93b