Correctly detect when nodes must be removed when moving content
With group eliding the composer can now delete a node directly
instead instead of a node always being surrounded by a group. This
triggered a latent issue detaching nodes from the tree to move them
to a new location; the chidl of a node that was directly removed
was not detected and its direct children were not removed causing
an exception to be thrown when the node was inserted into its
new location.

Fixes: 365802563
Test: new test, ./gradlew :composer:r:r:test
Change-Id: I048ef8315bb30c5b5d82b27f1a69f3db1892d9a6