Only assert not in suspending transaction if in compatibility mode.
Blocking DAO functions assert that they are not run on a suspending transaction to avoid deadlocks due to Android's thread local connection pool. This check should only be done if no driver is used as otherwise the connection pool manager in Room is Coroutine based and handles suspending transaction by default.

Test: BooksDaoTest
Change-Id: I2a986b29a0ef804f3d86d8cdb147c711a7ae1811