Refactor EditCommand to incorporate OutputTransformation
This CL is the first step for integrating OutputTransformation fully into IME commands. Until now the IME state synchronization with BasicTextField2 has been quite confusing with regards to OutputTransformation in terms of when the transformation was incorporated.

We are now trying to make the IME fully aware of OutputTransformation and its result as this is essentially what is presented to the user.

The next CL is going to finish the loop by adding transformed ranges to `NotifyImeListener`.

Test: existing tests for IME updates
Change-Id: I6a1009345d528f9914de168026d0ca5c7df6b5d9