replace non-sip addresses on sdks 26 & 27 for outgoing calls
after doing some investigating, it seems that inserting a non "tel"
scheme will cause OUTGOING VoIP calls to fail for sdks 26 & 27.  This
is a result of a missing change in the telecom platform that skips
validating VoIP call addresses.

In order to fix this in the jetpack layer, any outgoing call on sdk 26
and 27 will have the address changed to "sip:" + packageName if the
address does NOT have a "sip:" prefix.

Fixes: 372473795
Test: manual + unit
Change-Id: I7b2ebc57a2e384bcd431e44f988e4a110ee63728