Refactor PdfTestApp to Support Mutiple Scenarios
Currently, our PdfTestApp supports only single-PDF open scenario. With a growing user base for our
library, there is a need to extend the test app to handle multiple scenarios, making development, testing,
and validation more efficient. The goal is to enhance the test app so it can independently cover a wider
range of use cases for our library. This change introduces a reusable fragment setup that can be easily
adapted for future scenarios, consolidating common functionality in one place to reduce code duplication.

Bug: 368199470
Test: Validated changes on Pixel 8 pro
Change-Id: I0dbfe208ee7ad24da892d5427f7d21293b085b6a