Fix `kotlin.Unit` in KSP2
We convert kotlin.* types to Java equivalents in KSDeclaration.asJTypeName() by calling KSP's `Resolver.mapToJvmSignature()`. However, this function now (with 2.0.21-1.0.26-SNAPSHOT) maps `kotlin.Unit` to "V"(void) instead of "Lkotlin/Unit;" and this change breaks our code. This CL fixes that by not calling this function as we shouldn't convert `kotlin.Unit` for most cases in the first place.

There are only few cases we need `void` when getting Java typename:

1. `void` as the return type of a method in Java.
2. `kotlin.Unit` as the return type of a function in Kotlin directly (not a type variable).
3. `kotlin.Unit` as the return type of a Kotlin property setter.

And we use KspVoidType to cover these. From now on KspVoidType is only used when its Java typename should be 'void'. This also fixes an issue that sometimes we return 'void' when a parameter type is `kotlin.Unit`.

Test: XTypeNameTest.kt
Bug: 314151707
Change-Id: I08b4799ee9901b1e50cd61c84d4f8a6913af81e3