Improve "No Available Camera" Exception message with camera count
This change enhances the "No available camera can be found"
exception message by including the number of cameras
detected on the device.

When CameraX fails to find an available camera, it throws
an IllegalArgumentException with the message "No available camera
can be found." However, this message doesn't provide much
information about the underlying cause. It could be due to a
camera device failure during initialization, or some other
issue preventing camera access. This CL modifies the exception
message to include the number of cameras detected on the device.
This provides more context and helps in diagnosing the cause
of the issue.

Bug: 373284521
Test: Manually run test app on emulator
Change-Id: Idfd6aca9587253aa61b043e3fece1ebd20691054