Adjust CameraInternal shutdown behavior in UseCaseManager.close()
This CL adjusts the CameraInternal shutdown behavior in
UseCaseManager.close() such that it aligns with the same set of
behaviors in camera-camera2, while also making sure everything shuts
down properly.

We do this by isolating out the pieces necessary for shutdown (namely,
the CameraGraph and UseCaseCamera components should be closed), and only
do that without any additional UseCase invocations.

Bug: 375056795
Test: CameraControllerFragmentTest
Change-Id: Ie943b332008aec32088e739a67e03b5469295889