Fix lazy item animator target calculation when moving away
This change ensures items that are moving away towards the
end bound use the end of the last visible item as the
starting point of their target position calculation during
the lookahead pass.

The previous impl did not account of the size of the last
visible item. As a result, the animation may animate towards
the wrong direction when the last visible item is much larger
than the disappearing item.

Test: new test added
Change-Id: Iab83183122fd9185c630859d72df5b452c30ece6