room-paging refresh initial load fix
If items are removed then the initial key for a refresh may be close
enough to the end of the list that fewer than a screens-worth of items
may be loaded causing the UI to jump. Changed to handle similarly to the
case where the initial key larger than itemCount and load loadSize items
from the end.

Test: ./gradlew test connectedCheck
Fixes: b/389729367