CfW: Listen to the Clip events on the narrower scope rather than on the document (#2230)
When listening to Clip events on the document, we used to preventDefault
it, which is incorrect when there are other elements expecting Clip
events.

The Clip events target is not static:
If there is a backing textarea/input present and the Textfield is
focused, then it will be the events target. Otherwise, the `<canvas>`
itself will be the events target.

Partially fixes https://youtrack.jetbrains.com/issue/CMP-8492

## Testing
This should be tested by QA

## Release Notes
N/A