Fix ignored keyboard insets tests (#2346)
The new implementation of FocusTargetNode.focusRect() together with
access to the mutable state value causes endless content re-layout.
The OffsetToFocusedRect method assumes that getFocusedRect() should
return a non-observable rectangle of the focused element. To resolve
this issue, the code was modified to use the
`Snapshot.withoutReadObservation` method.

Fixes
https://youtrack.jetbrains.com/issue/CMP-8781/Fix-KeyboardInsetsTest-after-1.10.0-alpha01-merge

## Release Notes
N/A