Refactor buildSrc - Move `androidXComposeMultiplatform` extension to `org.jetbrains.androidx.build` (#2444)
Next steps after moving to the AOSP state:
- still keep `androidXComposeMultiplatform` instead of
`androidXMultiplatform`. I tried to get rid of it, but it requires a lot
of changes, even when we migrate to the AOSP buildSrc. More practical
path is to keep it until we have everything working
- extend from
[AndroidXMultiplatformExtension](https://github.com/androidx/androidx/blob/98817bea43540e93ecf9c9b7ce269c7c307ae1c2/buildSrc/private/src/main/kotlin/androidx/build/AndroidXMultiplatformExtension.kt#L75)
- remove some sourceSet configurations, reuse from the extension

## Testing
- [TeamCity
build](https://teamcity.jetbrains.com/buildConfiguration/JetBrainsPublicProjects_Compose_AllPersonalBuild/5538940)
- `./gradlew publishComposeJbToMavenLocal
-Pcompose.platforms=desktop,web` is successful
- wasm Demo works

## Release Notes
N/A