compose/ui/web: fix incorrect interpretation of `ontouchend` events (#2490)
In ComposeWindow.w3c.kt, the touch handler does not correctly handle the
`touchend` event. Indeed the logic selecting between `targetTouches` and
`changedTouches` is flawed. For `touchend` both should be taken into
account, also `changedTouches` contains the un-pressed pointer and
`targetTouches` contains the rest of the remaining pressed pointers.

The fix consists in removing the logic that compares the size of both
`targetTouches` and `changedTouches` to pick the largest one.
Instead, it checks the pointer event type to determine whether only
`targetTouches` or both `targetTouches` and `changedTouches` are needed
while correctly accounting which touches are pressed depending on the
source.

Fixes:
https://youtrack.jetbrains.com/issue/CMP-9030/wasm-canvas-incorrect-handling-of-multi-touch-input

## Testing
I have added a failing test in `GestureTest.kt` which is subsequently
fixed by the implementation commit. The test consists in using three
touches that are progressively pressed (1, 2, 3) and unpressed (3, 2,
1), logging and comparing the behavior of the pointerInput modifier with
the expected events.

## Release Notes
### Fixes - Web
- Fix incorrect interpretation of `ontouchend` events.