Fix navigation labels in mpp sample (#2610)
In mpp, titles are used as navigation labels that are supposed to be
unique. Currently, it's not possible to open material sample pages
because they are overwritten by material3 samples. These labels are not
supposed to be reused in UI, but for simplicity reasons we can just make
them unique

## Release Notes
N/A